Update Apache configuration
Set the UID and GID for the `www-data` user both to 1000. This allows
the Apache process to write to the host system during drupal install and
prevents. Assumes that developer has this UID on host system.

Enables rewrite module

Provides a default configuration to allow .htaccess files to be used to
rewrite requests. This allows use of clean URLs.